apps/api/pyproject.toml
[project]
name = "ledger-lift-api"
version = "0.1.0"
description = "FastAPI backend for Ledger Lift"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.0",
  "pydantic-settings==2.4.0",
  "SQLAlchemy==2.0.32",
  "psycopg2-binary==2.9.9",
  "boto3==1.34.131",
  "openpyxl==3.1.5",
  "alembic==1.13.1",
  "opentelemetry-distro==0.45b0",
  "opentelemetry-instrumentation-fastapi==0.45b0",
  "opentelemetry-instrumentation-sqlalchemy==0.45b0",
  "opentelemetry-exporter-otlp==1.24.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "httpx==0.27.0",
  "ruff==0.6.2"
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["E", "F", "I"]
